refactor: rest api를 고려하여 url 재적용
